backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
openpyxl==3.1.5
python-multipart==0.0.9
